+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The changelog below may be lengthy, but if you take the time to read it, I think
   - The game will also autosave before boss fights.
 - Added 2 warp doors in Sam's basement that lead to Bladesville and Telluria Castle Town.
 - Added a leaderboard for Fish Caught. It can be accessed by speaking to the fisherman in Telluria Castle Town.
+- Added an animation preview to the skill menu.
 - Added a golden First Cleric statue in the Frozen Labyrinth.
 - 4 new items:
   - Cleansing Scroll ★